backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile backend/requirements.in
#
aenum==3.1.15
    # via gremlinpython
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.10.11
    # via
    #   -r backend/requirements.in
    #   gremlinpython
aiosignal==1.3.1
    # via aiohttp
alembic==1.16.2
    # via -r backend/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via gremlinpython
attrs==24.2.0
    # via aiohttp
bcrypt==4.2.1
    # via -r backend/requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.2
    # via -r backend/requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
gremlinpython==3.7.3
    # via
    #   -r backend/requirements.in
    #   janusgraphpython
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   yarl
isodate==0.7.2
    # via gremlinpython
janusgraphpython==1.1.0
    # via -r backend/requirements.in
jinja2==3.1.6
    # via
    #   -r backend/requirements.in
    #   fastapi
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via gremlinpython
packaging==25.0
    # via -r backend/requirements.in
passlib==1.7.4
    # via -r backend/requirements.in
propcache==0.2.0
    # via yarl
psycopg2-binary==2.9.10
    # via -r backend/requirements.in
pydantic==2.9.2
    # via
    #   -r backend/requirements.in
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.9.1
    # via -r backend/requirements.in
pygments==2.18.0
    # via rich
pyjwt==2.10.1
    # via -r backend/requirements.in
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   -r backend/requirements.in
    #   fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.2
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.37
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.22
    # via -r backend/requirements.in
starlette==0.40.0
    # via fastapi
tornado==6.4.2
    # via -r backend/requirements.in
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic-settings
uvicorn[standard]==0.32.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
werkzeug==3.1.3
    # via -r backend/requirements.in
yarl==1.15.4
    # via aiohttp
